Snapshot finished
