+++ api/services/snapshot/store_getsnapshotbydate.go
@@ -2,6 +2,7 @@ package snapshot
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
@@ -13,15 +14,15 @@ func (s *PostgresSnapshotStore) GetSnapshotByDate(ctx context.Context, snapshotD
 
 	row := s.db.QueryRow(
 		c,
-		`
+		fmt.Sprintf(`
 			select 
-				* 
+				%s 
 			from 
 				snapshots
 			where 
 				user_id = $1
 				and snap_date = $2
-		`,
+		`, snapshotColumns),
 		userId,
 		snapshotDate,
 	)
