+++ api/services/snapshot/store_getsnapshotbyid.go
@@ -2,13 +2,11 @@ package snapshot
 
 import (
 	"context"
-	"database/sql"
 	"errors"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
-	"github.com/MicroFish91/portfolio-instruments-api/api/utils"
-	"github.com/jackc/pgx/v5"
 )
 
 func (s *PostgresSnapshotStore) GetSnapshotById(ctx context.Context, snapshotId, userId int) (types.Snapshot, []types.SnapshotValue, error) {
@@ -17,15 +15,15 @@ func (s *PostgresSnapshotStore) GetSnapshotById(ctx context.Context, snapshotId,
 
 	row := s.db.QueryRow(
 		c,
-		`
+		fmt.Sprintf(`
 			select 
-				* 
+				%s 
 			from 
 				snapshots
 			where 
 				user_id = $1
 				and snap_id = $2
-		`,
+		`, snapshotColumns),
 		userId,
 		snapshotId,
 	)
@@ -40,9 +38,9 @@ func (s *PostgresSnapshotStore) GetSnapshotById(ctx context.Context, snapshotId,
 
 	rows, err := s.db.Query(
 		c,
-		`
+		fmt.Sprintf(`
 			select 
-				* 
+				%s 
 			from 
 				snapshots_values
 			where 
@@ -51,7 +49,7 @@ func (s *PostgresSnapshotStore) GetSnapshotById(ctx context.Context, snapshotId,
 			order by 
 				account_id ASC, 
 				holding_id ASC
-		`,
+		`, snapshotValueColumns),
 		userId,
 		snapshotId,
 	)
@@ -67,56 +65,4 @@ func (s *PostgresSnapshotStore) GetSnapshotById(ctx context.Context, snapshotId,
 		return types.Snapshot{}, nil, err
 	}
 	return snapshot, snapshotValues, nil
-}
-
-func (s *PostgresSnapshotStore) parseRowIntoSnapshot(row pgx.Row) (types.Snapshot, error) {
-	var snap types.Snapshot
-	var benchmark_id, rebalance_threshold_pct sql.NullInt64
-
-	err := row.Scan(
-		&snap.Snap_id,
-		&snap.Description,
-		&snap.Snap_date,
-		&snap.Total,
-		&snap.Weighted_er_pct,
-		&benchmark_id,
-		&snap.User_id,
-		&snap.Created_at,
-		&snap.Updated_at,
-		&rebalance_threshold_pct,
-	)
-
-	if err != nil {
-		return types.Snapshot{}, err
-	}
-
-	snap.Benchmark_id = utils.ConvertNullIntToInt(benchmark_id)
-	snap.Rebalance_threshold_pct = utils.ConvertNullIntToInt(rebalance_threshold_pct)
-
-	return snap, nil
-}
-
-func (s *PostgresSnapshotStore) parseRowsIntoSnapshotValues(rows pgx.Rows) ([]types.SnapshotValue, error) {
-	var snapshotValues []types.SnapshotValue
-	for rows.Next() {
-		var sv types.SnapshotValue
-		err := rows.Scan(
-			&sv.Snap_val_id,
-			&sv.Snap_id,
-			&sv.Account_id,
-			&sv.Holding_id,
-			&sv.Total,
-			&sv.Skip_rebalance,
-			&sv.User_id,
-			&sv.Created_at,
-			&sv.Updated_at,
-		)
-
-		if err != nil {
-			return nil, err
-		}
-		snapshotValues = append(snapshotValues, sv)
-	}
-
-	return snapshotValues, nil
 }
