--- api/stores.txt
@@ -1,13 +1,0 @@
-./api/services/snapshot/store_groupbymaturationdate.go
-./api/services/snapshot/store_getsnapshotbydate.go
-./api/services/snapshot/store_getsnapshotbyid.go
-./api/services/snapshot/store_getsnapshottotal.go
-./api/services/snapshot/store_deletesnapshot.go
-./api/services/snapshot/store_createsnapshot.go
-./api/services/snapshot/store_groupbyholding.go
-./api/services/snapshot/store_groupbyliquidity.go
-./api/services/snapshot/store_updatesnapshot.go
-./api/services/snapshot/store_refreshweighteder.go
-./api/services/snapshot/store_getsnapshots.go
-./api/services/snapshot/store_refreshtotal.go
-./api/services/snapshot/store_groupbyaccount.go
